Don't use rawValues if raw scanner fails to parse the ini file
In some cases (such as when the ini file contains multi-line strings),
parse_ini_file with the INI_SCANNER_RAW flag causes the parse to fail.
This in turn leads to a lot of noise in error logs because the
subsequent call to decode throws a lot of E_NOTICEs. If the parse fails,
just use what was processed by parse_ini_file directly instead.

See matomo-org/matomo#15689 for additional context.